check feature parity for feature engineering
